refactor(screens): reorganize screen directories
