+++ app/src/navigation/account.web.ts
@@ -1,0 +1,38 @@
+// SPDX-FileCopyrightText: 2025 Social Connect Labs, Inc.
+// SPDX-License-Identifier: BUSL-1.1
+// NOTE: Converts to Apache-2.0 on 2029-06-11 per LICENSE.
+
+import type { NativeStackNavigationOptions } from '@react-navigation/native-stack';
+
+import DocumentDataNotFoundScreen from '@/screens/account/recovery/DocumentDataNotFoundScreen';
+import SettingsScreen from '@/screens/account/settings/SettingsScreen';
+import { black, white } from '@/utils/colors';
+
+const accountScreens = {
+  DocumentDataNotFound: {
+    screen: DocumentDataNotFoundScreen,
+    options: {
+      headerShown: false,
+      gestureEnabled: false,
+      animation: 'slide_from_bottom',
+    } as NativeStackNavigationOptions,
+  },
+  Settings: {
+    screen: SettingsScreen,
+    options: {
+      animation: 'slide_from_bottom',
+      title: 'Settings',
+      headerStyle: {
+        backgroundColor: white,
+      },
+      headerTitleStyle: {
+        color: black,
+      },
+    } as NativeStackNavigationOptions,
+    config: {
+      screens: {},
+    },
+  },
+};
+
+export default accountScreens;
