--- app/src/navigation/recovery.ts
@@ -1,67 +1,0 @@
-// SPDX-FileCopyrightText: 2025 Social Connect Labs, Inc.
-// SPDX-License-Identifier: BUSL-1.1
-// NOTE: Converts to Apache-2.0 on 2029-06-11 per LICENSE.
-
-import type { NativeStackNavigationOptions } from '@react-navigation/native-stack';
-
-import AccountRecoveryChoiceScreen from '@/screens/recovery/AccountRecoveryChoiceScreen';
-import AccountRecoveryScreen from '@/screens/recovery/AccountRecoveryScreen';
-import AccountVerifiedSuccessScreen from '@/screens/recovery/AccountVerifiedSuccessScreen';
-import DocumentDataNotFound from '@/screens/recovery/DocumentDataNotFoundScreen';
-import RecoverWithPhraseScreen from '@/screens/recovery/RecoverWithPhraseScreen';
-import SaveRecoveryPhraseScreen from '@/screens/recovery/SaveRecoveryPhraseScreen';
-import { black, slate300 } from '@/utils/colors';
-
-const recoveryScreens = {
-  AccountRecovery: {
-    screen: AccountRecoveryScreen,
-    options: {
-      headerShown: false,
-    } as NativeStackNavigationOptions,
-  },
-  AccountRecoveryChoice: {
-    screen: AccountRecoveryChoiceScreen,
-    options: {
-      headerShown: false,
-    } as NativeStackNavigationOptions,
-  },
-  AccountVerifiedSuccess: {
-    screen: AccountVerifiedSuccessScreen,
-    options: {
-      headerShown: false,
-      animation: 'slide_from_bottom',
-    } as NativeStackNavigationOptions,
-  },
-  DocumentDataNotFound: {
-    screen: DocumentDataNotFound,
-    options: {
-      headerShown: false,
-      gestureEnabled: false,
-      animation: 'slide_from_bottom',
-      // presentation: 'modal',
-    } as NativeStackNavigationOptions,
-  },
-  RecoverWithPhrase: {
-    screen: RecoverWithPhraseScreen,
-    options: {
-      headerTintColor: black,
-      title: 'Enter Recovery Phrase',
-      headerStyle: {
-        backgroundColor: black,
-      },
-      headerTitleStyle: {
-        color: slate300,
-      },
-      headerBackTitle: 'close',
-    } as NativeStackNavigationOptions,
-  },
-  SaveRecoveryPhrase: {
-    screen: SaveRecoveryPhraseScreen,
-    options: {
-      headerShown: false,
-      animation: 'slide_from_bottom',
-    } as NativeStackNavigationOptions,
-  },
-};
-
-export default recoveryScreens;
