--- app/src/navigation/settings.ts
@@ -1,75 +1,0 @@
-// SPDX-FileCopyrightText: 2025 Social Connect Labs, Inc.
-// SPDX-License-Identifier: BUSL-1.1
-// NOTE: Converts to Apache-2.0 on 2029-06-11 per LICENSE.
-
-import type { NativeStackNavigationOptions } from '@react-navigation/native-stack';
-
-import CloudBackupScreen from '@/screens/settings/CloudBackupScreen';
-import DocumentDataInfoScreen from '@/screens/settings/DocumentDataInfoScreen';
-import ManageDocumentsScreen from '@/screens/settings/ManageDocumentsScreen';
-import SettingsScreen from '@/screens/settings/SettingsScreen';
-import ShowRecoveryPhraseScreen from '@/screens/settings/ShowRecoveryPhraseScreen';
-import { black, slate300, white } from '@/utils/colors';
-
-const settingsScreens = {
-  CloudBackupSettings: {
-    screen: CloudBackupScreen,
-    options: {
-      title: 'Cloud backup',
-      headerStyle: {
-        backgroundColor: black,
-      },
-      headerTitleStyle: {
-        color: slate300,
-      },
-    } as NativeStackNavigationOptions,
-  },
-  ManageDocuments: {
-    screen: ManageDocumentsScreen,
-    options: {
-      title: 'Manage Documents',
-      headerStyle: {
-        backgroundColor: white,
-      },
-      headerTitleStyle: {
-        color: black,
-      },
-    } as NativeStackNavigationOptions,
-  },
-  DocumentDataInfo: {
-    screen: DocumentDataInfoScreen,
-    options: {
-      title: 'Document Data Info',
-      headerStyle: {
-        backgroundColor: white,
-      },
-    } as NativeStackNavigationOptions,
-  },
-  Settings: {
-    screen: SettingsScreen,
-    options: {
-      animation: 'slide_from_bottom',
-      title: 'Settings',
-      headerStyle: {
-        backgroundColor: white,
-      },
-      headerTitleStyle: {
-        color: black,
-      },
-    } as NativeStackNavigationOptions,
-    config: {
-      screens: {},
-    },
-  },
-  ShowRecoveryPhrase: {
-    screen: ShowRecoveryPhraseScreen,
-    options: {
-      title: 'Recovery Phrase',
-      headerStyle: {
-        backgroundColor: white,
-      },
-    } as NativeStackNavigationOptions,
-  },
-};
-
-export default settingsScreens;
