+++ app/src/screens/account/settings/ShowRecoveryPhraseScreen.tsx
