+++ app/src/screens/documents/management/IdDetailsScreen.tsx
