+++ app/src/screens/documents/scanning/DocumentCameraScreen.tsx
