+++ app/src/screens/documents/scanning/DocumentCameraTroubleScreen.tsx
