+++ app/src/screens/documents/scanning/DocumentNFCMethodSelectionScreen.tsx
