+++ app/src/screens/documents/scanning/DocumentNFCScanScreen.tsx
