+++ app/src/screens/documents/scanning/DocumentNFCTroubleScreen.tsx
