+++ app/src/screens/onboarding/AccountVerifiedSuccessScreen.tsx
@@ -14,7 +14,7 @@ import { PrimaryButton } from '@/components/buttons/PrimaryButton';
 import Description from '@/components/typography/Description';
 import { Title } from '@/components/typography/Title';
 import { ExpandableBottomLayout } from '@/layouts/ExpandableBottomLayout';
-import { styles } from '@/screens/prove/ProofRequestStatusScreen';
+import { styles } from '@/screens/verification/ProofRequestStatusScreen';
 import { black, white } from '@/utils/colors';
 import { buttonTap } from '@/utils/haptic';
