+++ app/src/screens/verification/ProveScreen.tsx
