+++ app/src/screens/verification/QRCodeTroubleScreen.tsx
