+++ app/src/screens/verification/QRCodeViewFinderScreen.tsx
