+++ app/vite.config.ts
@@ -173,11 +173,11 @@ export default defineConfig({
           'vendor-state-zustand': ['zustand'],
 
           // Screen-specific chunks - more granular
-          'screens-document-core': ['./src/navigation/document.ts'],
+          'screens-document-core': ['./src/navigation/documents.ts'],
           'screens-passport-nfc': ['./src/utils/nfcScanner.ts'],
 
           // Proving - split into even smaller chunks
-          'screens-prove-core': ['./src/navigation/prove.ts'],
+          'screens-prove-core': ['./src/navigation/verification.ts'],
           'screens-prove-validation-core': [
             './src/utils/proving/validateDocument.ts',
           ],
@@ -192,8 +192,8 @@ export default defineConfig({
           ],
 
           // Other screens
-          'screens-settings': ['./src/navigation/settings.ts'],
-          'screens-recovery': ['./src/navigation/recovery.ts'],
+          'screens-settings': ['./src/navigation/account.ts'],
+          'screens-recovery': ['./src/navigation/account.ts'],
           'screens-dev': ['./src/navigation/devTools.ts'],
         },
       },
